backend-oss/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
pydantic[email]==2.10.4
python-multipart==0.0.9
jinja2==3.1.4
